SCP-2933 - Move Marlowe types to a separate module
