+++ marlowe/marlowe.cabal
@@ -77,8 +77,10 @@ library
     Language.Marlowe.Deserialisation
     Language.Marlowe.Extended
     Language.Marlowe.Semantics
+    Language.Marlowe.SemanticsTypes
     Language.Marlowe.Client
     Language.Marlowe.Util
+    Language.Marlowe.ParserUtil
     Language.Marlowe.Scripts
     Language.Marlowe.Serialisation
     Language.Marlowe.Pretty
