+++ marlowe/src/Language/Marlowe.hs
@@ -1,5 +1,6 @@
 module Language.Marlowe
     ( module Language.Marlowe.Semantics
+    , module Language.Marlowe.SemanticsTypes
     , module Language.Marlowe.Client
     , module Language.Marlowe.Util
     , module Language.Marlowe.Pretty
@@ -13,7 +14,8 @@ where
 import           Language.Marlowe.Client
 import           Language.Marlowe.Pretty
 import           Language.Marlowe.Semantics
+import           Language.Marlowe.SemanticsTypes
 import           Language.Marlowe.Util
-import           Ledger                     (Slot (..))
-import           Ledger.Ada                 (adaSymbol, adaToken)
+import           Ledger                          (Slot (..))
+import           Ledger.Ada                      (adaSymbol, adaToken)
 import           PlutusTx.Ratio
