+++ marlowe/src/Language/Marlowe/Analysis/FSSemantics.hs
@@ -5,23 +5,24 @@
 {-# OPTIONS_GHC -Wno-unused-imports #-}
 module Language.Marlowe.Analysis.FSSemantics where
 
-import           Data.List                  (foldl', genericIndex)
-import           Data.Map.Strict            (Map)
-import qualified Data.Map.Strict            as M
-import           Data.Maybe                 (isNothing)
+import           Data.List                       (foldl', genericIndex)
+import           Data.Map.Strict                 (Map)
+import qualified Data.Map.Strict                 as M
+import           Data.Maybe                      (isNothing)
 import           Data.SBV
-import qualified Data.SBV.Either            as SE
-import           Data.SBV.Internals         (SMTModel (..))
-import qualified Data.SBV.List              as SL
-import qualified Data.SBV.Maybe             as SM
-import qualified Data.SBV.Tuple             as ST
-import           Data.Set                   (Set)
-import qualified Data.Set                   as S
+import qualified Data.SBV.Either                 as SE
+import           Data.SBV.Internals              (SMTModel (..))
+import qualified Data.SBV.List                   as SL
+import qualified Data.SBV.Maybe                  as SM
+import qualified Data.SBV.Tuple                  as ST
+import           Data.Set                        (Set)
+import qualified Data.Set                        as S
 import           Language.Marlowe.Semantics
-import           Ledger                     (Slot (..))
-import qualified PlutusTx.AssocMap          as AssocMap
-import qualified PlutusTx.Prelude           as P
-import qualified PlutusTx.Ratio             as P
+import           Language.Marlowe.SemanticsTypes
+import           Ledger                          (Slot (..))
+import qualified PlutusTx.AssocMap               as AssocMap
+import qualified PlutusTx.Prelude                as P
+import qualified PlutusTx.Ratio                  as P
 
 ---------------------------------------------------
 -- Static analysis logic and symbolic operations --
