+++ marlowe/src/Language/Marlowe/Util.hs
@@ -1,17 +1,17 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 module Language.Marlowe.Util where
-import           Data.List                  (foldl')
-import           Data.Map.Strict            (Map)
-import qualified Data.Map.Strict            as Map
-import           Data.Set                   (Set)
-import qualified Data.Set                   as Set
+import           Data.List                       (foldl')
+import           Data.Map.Strict                 (Map)
+import qualified Data.Map.Strict                 as Map
+import           Data.Set                        (Set)
+import qualified Data.Set                        as Set
 import           Data.String
-
 import           Language.Marlowe.Semantics
-import           Ledger.Ada                 (adaSymbol, adaToken)
-import qualified Ledger.Value               as Val
-import qualified PlutusTx.Prelude           as P
+import           Language.Marlowe.SemanticsTypes
+import           Ledger.Ada                      (adaSymbol, adaToken)
+import qualified Ledger.Value                    as Val
+import qualified PlutusTx.Prelude                as P
 
 instance IsString Party where
     fromString s = Role (fromString s)
