+++ nix/pkgs/haskell/materialized-windows/.plan.nix/marlowe.nix
@@ -62,8 +62,10 @@
           "Language/Marlowe/Deserialisation"
           "Language/Marlowe/Extended"
           "Language/Marlowe/Semantics"
+          "Language/Marlowe/SemanticsTypes"
           "Language/Marlowe/Client"
           "Language/Marlowe/Util"
+          "Language/Marlowe/ParserUtil"
           "Language/Marlowe/Scripts"
           "Language/Marlowe/Serialisation"
           "Language/Marlowe/Pretty"
